comments

+++ src/emit/pretty_printer/doc_impl.cpp
@@ -169,7 +169,7 @@ auto resolveAlignment(const DocPtr &doc) -> DocPtr
         using T = std::decay_t<decltype(node)>;
         if constexpr (std::is_same_v<T, AlignText>) {
             const int max_width = max_widths.at(node.level);
-            const int padding = max_width - node.content.length();
+            const int padding = max_width - static_cast<int>(node.content.length());
             return makeText(node.content + std::string(padding, ' '));
         } else {
             return std::make_shared<DocImpl>(node);
